+++ src/controllers/QueueController.php
@@ -14,9 +14,16 @@ class QueueController extends Controller
     {
         $params = \Craft::$app->request->bodyParams;
 
+        $phone = $params['phone'];
+
+        // if landcode hidden input field is set, append this to phone number unless phone number starts with +
+        if(isset($params['country_code']) && substr($phone, 0, strlen('+')) !== '+'){
+            $phone = '+' . $params['country_code'] . $phone;
+        }
+
         $guest = (new Guest())
             ->setEmail($params['email'])
-            ->setPhone($params['phone'])
+            ->setPhone($phone)
             ->setName($params['name']);
 
         if($guest->validate()){